feat: 🎸 Home page is calling the store and functioning normally
The home page is now using data from the store, and there was a change
in the Header's styling to improve its accessibility. The
IntroductionBox component, which will be used in the introduction of the
pages, was also created. I made a small change to the dogPostSlice, and
now the AsyncThunk gettingDogsImage accepts a variable that allows
controlling how many images will be returned.